chore: split solid
